+++ internal/autodev/CHANGELOG.md
@@ -41,3 +41,9 @@
 
 ## [0.4.9+1781661144] - 2026-06-17
 - Test integration task
+
+## [0.4.9+1781667645] - 2026-06-17
+- Test integration task
+
+## [0.4.9+1781667780] - 2026-06-17
+- Test integration task